Update Debian Install Script To Support Debian 10
Support for installing pwsh on Debian 10 was added in #11236, but the
install script for Debian was not updated. This resulted in an error
when running the install script that claimed the distro was not
supported.

This change simply adds `10` to the collection of supported distro's.

I have tested that the script will now successfully install powershell
preview 7 on a Debian 10 machine.

Fix #11539