Handle cases where CustomEvent was not initially sent
Since the CustomEvent is sent only on startup, we need a way to send
the startup event in those cases where PowerShell is being hosted
(not the console). This will check to see if we have already sent the
custom event for startup and send if it hasn't. This will enable us
to track some of the hosted scenarios like Jupyter.