:wrench: Improve detection of mutable value types
Currently, we have some flawed assumptions around what defines a mutable
value type. Rather than make a more complicated series of checks and
making assumptions, we can simply look at whether the type is defined
with the IsReadOnlyAttribute.

As a fallback, it's much safer and less complex to enumerate public and
private fields on the type, and verify whether they're all read-only.

Resolves #12430.