Bind self-referential generic type parameters late
By supplying the interface last up front when defining a new class, we rob
ourselves the opportunity to resolve the class in question as a generic type
parameter. This makes the following implementation - a completely valid type
definition - impossible in PowerShell:

    class SortableThing : IComparable[SortableThing]
    {
        [int]Value

        [int]
        CompareTo([SortableThing]$obj)
        {
            return $this.Value.CompareTo($obj.Value)
        }
    }

This commit introduces an InterfaceExpression helper to bind the current
typebuilder to any matching type parameter in it's own interface list _after_
initial definition.