:sparkles: Let events be accessed by name
This gives visibility into events so that users can target them.
It also vastly simplifies the necessary handling in the compiler/binders
as we can use a runtime conditional expression to determine if the
target of the += / -= op is a PSEvent and maintain some sanity in the
binders and compiler logic needed here.

Completion logic also needed to be enhanced to provide completion for
event names.