:bug: Resolve issue in history input handling
This resolves the fragile conversions in the history input handling.

Essentially, anything taking input in this manner
(values nested in a hashtable / other loosely-typed property bag)
need to be careful to call conversion APIs or always be careful to check
for and retrieve the BaseObject if the input happens to be PSObject.

This was previously "fine" largely by accident, and would only have
been noticed previously for unrealistically simple structs which were
mistakenly not rewrapped as PSObject.

I additionally reframed the other conversions in this method to follow
a common pattern and use existing functionality in LanguagePrimitives.