Start transcript fix non append encoding
When the -Append file has not been set by the user, the file is emptied using WriteAllText.
This method seems to keep the encoding of the file if it existed prior.
We don't want this behavior and want to be
sure that the UTF8NoBom is used.

It calls the override with the encoding set.