Merged PR 17522: Add-type NoLanguage mode vuln fix
Normally, the Add-Type cmdlet is disallowed when PowerShell is run on a locked down system.  However, an admin on a locked down system can create a NoLanguage remoting endpoint, and that endpoint configuration will allow the Add-Type cmdlet.  The admin can create a remote connection to the NoLanguage endpoint on the same machine, and use Add-Type cmdlet to compile and run arbitrary C# code.

This fix disallows the use of Add-Type cmdlet in a NoLanguage mode PowerShell session, on a locked down machine.

Cherry picked from !17521