Retrieve more info with Get-AuthenticodeSignature
Adds the Signatures property to the Signature output object of
Get-AuthenticodeSignature and Set-AuthenticodeSignature. This new object
is an array of signature detail objects that contains the signing and
digest algorithm as well as the timestamp information of the signature.
It does this for all the entries in the authenticode block and not just
the first entry