Check to be sure we're calling powershell before encoding a scriptblock.
If we are not calling powershell/pwsh a scriptblock will be passed as is rather than base64 encoded.
This will alleviate the need for quoting commands like:
`msiexec /p "msipatch.msp;msipatch2.msp" /n {00000001-0002-0000-0000-624474736554} /qb`.
Arguments will be passed as follows:

```powershell
PS> testexe -echoargs a b {get-date}
Arg 0 is <a>
Arg 1 is <b>
Arg 2 is <{get-date}>
```

This is a breaking change, but any current work-arounds should not need alteration.