Add additional platform check for Windows before launching native app.
This change ensures that we do not get the cursor position in preparation for the BufferCell call since it is only supported on Windows.
On non-Windows machines this might also cause problems for applications such as xpect in which pwsh runs in a session connected by a pty.
This change should address the misbehavior described in issue #18874.