Fix error view for pipeline enumeration exceptions
When the pipeline was invoking a `IEnumerator.MoveNext` implementation
that threw an exception, that exception was failing to render. This was
occuring because the formatter was trying to test for
`TargetObject.Line`. The `TargetObject` in this case was the enumerator,
so the engine attempted member access enumeration which threw the same
exception.

This change fixes that by explicitly testing the type of `TargetObject`
prior to attempting member access.