Fix `Get-Service` non-terminating error message to include category (#20276)
et-Service cmdlet has a common method for emitting non-terminating error message. Unfortunately, it assumed that if there isn't an inner exception message to just leave the content blank when formatting the error string. Fix here is that if there isn't an inner exception message, we use the string value of the error category instead of an empty string.