WIP: Maintain http persistence where possible
A default web session is created whenever the user doesn't
provide a session or request it to be saved somewhere.

When the default session is re-used, any changes made to it
are reset so the request will behave predictably. The reset
process tries to ensure maximum opportunities for persistent
http requests by avoiding recreating the http client when
subsequent invocations use compatible parameters.

The default web session is saved as a global session state
variable called __DefaultWebSession. This approach is
used so that each run space will get its own default web
session to avoid problems when running requests in parallel.