Set-Acl: Do not fail on untranslatable SID
Fix up error when attempting to set a SecurityDescriptor that contains a
SACL, Owner, or Group entry with a SecurityIdentifier that cannot be
translated. For example the existing SD or SD to set owner has a
SecurityIdentifier set to a user that doesn't exist on the current
machine or is a domain account where the host isn't domain joined.