Fix Get-Error serialization of array values (#21085)
Ensure Get-Error does not hang when attempting to serialize an exception
that contains a property whose type is an array of System.Type
instances. Also ensures that primitive types like System.Int32,
System.Boolean as well as System.String is formatted as a string rather
than an object with properties.