Get-Process: Remove admin for IncludeUserName
Removes the administrator requirements checked on Get-Process
-IncludeUserName as it can be used to get the username on processes you
currently own. The check was unecessary and the value can be set to null
on processes the current user does not have access to.

This also changes the logic for translating the SecurityIdentifier in
the access token to use the builtin .NET class which handles more
scenarios (like Font Driver Host users), and is simpler.