Start-Process: Fix -Credential error as non-admin
When using Start-Process -Credential in a non-administrator context with
a different user's credentials the code currently throws an exception
when attempting to get the new process' handle. This change treats this
error as a warning only if -PassThru was specified as it only affects
certain members on the returned object like ExitCode but others like
WaitForExit() still work fine.