Improve Array BinaryAdd Method for object[]
Use Array.Resize when doing += on an array type. This is more efficient
than the default BinaryAdd done on other enumerable types. As adding to
an array is a common operation done by people new to PowerShell this
should reduce some of the issues which come with that type of code.