Fix Hyper-V Remoting when using remoting import
When on a Windows version that cannot load the Hyper-V module the State
property will be of the string type. This change ensures that PowerShell
can still parse this value to the VMState type as expected when the
module was loaded natively. This should allow PSRemoting cmdlets that
call Get-VM to properly get the VM State property.