WebCmdlets check if XML declaration or HTML meta encoding value make sense
WebCmdlets looks for XML declaration or HTML meta encoding value in the default encoding (UTF-8 at the time of the commit).
If the parsed value is not readable anymore after encoding change it means the encoding value is false.
This change is for cases where the declared encoding is UTF-7/16/32 which are incompatible with UTF-8. These encoding would have a BOM header.
It adds a check that dynamically look for the XML declaration or HTML meta encoding after encoding change.