Get-Service Ignore common errors
Ignore common errors when attempting to retrieve some properties for a
service like the Description, UserName, StartupType. These properties
are not critical for using a service object or to retrieve other
information like the name and whether a service exists or not and by
emitting an error record it could fail a script if it did not expect an
error.

Instead if the service failed to retrieve the property due to
permissions or other environmental problems, it will set the affected
properties to $null.