Fix config.json PSModulePath path with sep char
Fixes the parsing of the powershell.config.json PSModulePath entry when
it contains the path separator char. Instead of crashing this is able to
handle entries with those characters and treat them as multiple entries.

This also rewrites the logic when building the process PSModulePath to
simplify the way entries were added and deal with canonicalisation of
the path entries.