Fix debug tracing error with magic extents
Fixes the error when PowerShell attempts to print the debug trace line
for functions generated with an invalid extent. For example PowerShell
creates an extent with a column and line of 0 for things like the
default class constructors. This is an issue as the trace debug line
always subtracts by 1 to make the value start at 0 rather than 1.

It also adds the DebuggerHidden attribute to the class default
constructor to avoid confusing end users from seeing a trace line for
something they did not write.