Trusted Publisher Checks for Azure Trusted Signing
Add support for checking the Azure Trusted Signing publisher identifier
alongside the thumbprint. This check will verify whether the unique
Azure TS OID present in the EKU is in any certificate in the
TrustedPublishers store. It will also check whether the OID is present
in the untrusted store making it possible to distrust all certificates
for that Azure identity.