New-PSSessionConfigurationFile - value handling
Improves the parameter handling to properly unwrap PSObject wrapped
string or Hashtable values when encountered in object[] based
parameters. The existing code just checked for the string or Hashtable
types without considering whether they were wrapped by a PSObject.

This also improves the error message to include the correct member name
on an error as before it was hardcoded to `ModulesToImport`.